infra/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "service-cdk"
version = "1.0"
description = "CDK service code"
authors = [
    { name = "Ran Isenberg", email = "ran.isenberg@ranthebuilder.cloud" }
]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3.12"
]
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "aws-cdk.aws-lambda-python-alpha>=2.170.0",
    "constructs>=10.0.0",
    "boto3>=1.35.0",
    "cdk-nag>=2.28.0",
    "cdk-monitoring-constructs>=8.0.0",
    "aws-lambda-powertools>=2.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["contrib", "docs", "tests"]

[tool.setuptools.package-data]
"*" = ["*.json"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
